dropa/requirements.txt
django
djangorestframework
psycopg2-binary
pyotp
pandas
numpy
scikit-learn
prophet
joblib
lightgbm
xgboost
